+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 - added French translation ([#43](https://github.com/AlmostReliable/energymeter/pull/43))
+- added CC: Tweaked integration ([#42](https://github.com/AlmostReliable/energymeter/pull/42))
+- fixed Energy Meter not storing changes when only the total energy changes
 
 ## [0.4.1] - 2026-05-16